+++ lib/src/widgets/provider_builder.dart
@@ -2,6 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
+import '../../riverpod_extension.dart';
+
 /// A widget helper for displaying [StreamProvider] data
 class ProviderStreamBuilder<T> extends HookWidget {
   /// A widget helper for displaying [StreamProvider] data
@@ -49,6 +51,53 @@ class ProviderFutureBuilder<T> extends HookWidget {
         data: builder,
         error: (error, stackTrace) => builder(null),
         loading: () => loading ?? builder(null),
+      );
+}
+
+/// A widget helper for displaying [FutureProvider] data
+class ProviderHttpFutureBuilder<T> extends HookWidget {
+  /// A widget helper for displaying [FutureProvider] data
+  const ProviderHttpFutureBuilder({
+    required this.provider,
+    required this.data,
+    required this.loading,
+    required this.error,
+    Key? key,
+  }) : super(key: key);
+
+  /// The provider for obtaining the future
+  final AutoDisposeFutureProvider<HttpResponseState<T>> provider;
+
+  /// Widget to display when data is retrieved from the future
+  final Widget Function(T value) data;
+
+  /// Widget to display when data is loading from the future
+  final Widget Function() loading;
+
+  /// Widget to display when there is an error when retrieving data from the
+  /// future
+  final Widget Function(
+    Object error,
+    StackTrace? stackTrace,
+  ) error;
+
+  @override
+  Widget build(BuildContext context) => useProvider(provider).when(
+        data: (value1) => value1.when(
+          (value2) {
+            if (value2 == null) {
+              return loading();
+            }
+            return data(value2);
+          },
+          loading: loading,
+          error: (exception) => error(
+            exception,
+            null,
+          ),
+        ),
+        loading: loading,
+        error: error,
       );
 }