+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+
+* Add ProviderHttpFutureBuilder for HTTP requests
+* Add pushRoute for providers
+* Add retries to requests
+
 ## 1.0.3
 
 * Remove firebase dependency